Finish factoring out remaining storage & sync concerns from Location to Sync.
