+++ Sources/Location/LocationAssembler.swift
@@ -32,26 +32,5 @@ public func assemble(container: Container) {
         container.register(RegionManager.self) { resolver in
             return resolver.resolve(LocationManager.self)!
         }
-        
-        container.register(SyncParticipant.self, name: "location.beacons") { resolver in
-            return BeaconsSyncParticipant(
-                context: resolver.resolve(NSManagedObjectContext.self, name: "location.backgroundContext")!,
-                userDefaults: UserDefaults.standard
-            )
-        }
-        
-        container.register(SyncParticipant.self, name: "location.geofences") { resolver in
-            return GeofencesSyncParticipant(
-                context: resolver.resolve(NSManagedObjectContext.self, name: "location.backgroundContext")!,
-                userDefaults: UserDefaults.standard
-            )
-        }
-    }
-    
-    public func containerDidAssemble(resolver: Resolver) {
-        resolver.resolve(SyncCoordinator.self)!.participants.append(contentsOf: [
-            resolver.resolve(SyncParticipant.self, name: "location.beacons")!,
-            resolver.resolve(SyncParticipant.self, name: "location.geofences")!
-        ])
     }
 }
