+++ Sources/Sync/SyncCoordinator/Participants/BeaconsSyncParticipant.swift
@@ -1,6 +1,6 @@
 //
 //  BeaconsSyncParticipant.swift
-//  RoverLocation
+//  RoverSync
 //
 //  Created by Sean Rucker on 2018-09-06.
 //  Copyright © 2018 Rover Labs Inc. All rights reserved.
