+++ Sources/Sync/SyncCoordinator/Participants/GeofencesSyncParticipant.swift
@@ -1,6 +1,6 @@
 //
 //  GeofencesSyncParticipant.swift
-//  RoverLocation
+//  RoverSync
 //
 //  Created by Sean Rucker on 2018-08-29.
 //  Copyright © 2018 Rover Labs Inc. All rights reserved.
